+++ src/App.tsx
@@ -6,6 +6,7 @@ import DatesPage from './pages/DatesPage'
 import CurrencyPage from './pages/CurrencyPage'
 import LocaleNamesPage from './pages/LocaleNamesPage'
 import PluralRulesPage from './pages/PluralRulesPage'
+import SegmentationPage from './pages/SegmentationPage'
 
 function App() {
   return (
@@ -18,6 +19,7 @@ function App() {
           <Route path="currency" element={<CurrencyPage />} />
           <Route path="locale-names" element={<LocaleNamesPage />} />
           <Route path="plural-rules" element={<PluralRulesPage />} />
+          <Route path="segmentation" element={<SegmentationPage />} />
         </Route>
       </Routes>
     </BrowserRouter>